refs #12228 split mantid and mantiddev into 2 parsers
